[server] Change LOG_DEBUG to LOG_INFO
We have a -D flag, which turns the extra output on and off. The way we
use it, it is an equivalent of using setlogmask (which we don't use).

Since -D allows to flip debugging on and off, changes to rsyslog.conf
are not necessary. So, to capture the extra output in /var/log/messages,
change all debugging messages to LOG_INFO priority.

The alternative would be to kill -D entirely. But it's easier to flip
a flag than to muck with rsyslog.conf, so let's do it this way.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>