tabled: Use config file instead of most options
Some of the new configuration options that we need are not convenient
to express on the command line, notably the listening on the SSL sockets.
So let's add a configuration file capability, to be used later.

This patch is essentially ripped off chunkd, but I just had to tweak it:
plug some small memory leaks, detect more syntax errors.

Note that "make distcheck" still passes. The generated configuration
is removed, so it does not end in the tarball.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>