tabled: Remember object ID counter
Currently we verify that there's no object ID conflict by using the
error from open(2) on the local backing store. This does not work
as well when the backing store is chunkd. Firstly, we add a round
trip to server, and secondly it is advantageous not to have a duplicate
verification in the non-replicated store.

I pondered the way to manage the object ID, which would not rely on
randomness, and it seems that updating a database of IDs on every
object creation would be excessive. So, instead, remember the current
object ID counter. Unlike a normal database of IDs, we only update
the permanent storage once in 500 transactions. We store the ID in
a fake database so that replication picks this up in the future.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>