tabled: add flock for PID file
I'm tired of removing the PID file in case of a crash, so this patch
implements a way to determine if it's stale.

We do not use the usual trick of kill(pid, 0), because that tends
to fail sometimes when a bunch of various daemons start on boot in
modern parallel-booting systems.

Note that flock works in NFS-mounted roots, in theory at least.

The pid_fd is not used for anything tangible. It only demonstrates
that we want to keep the PID file open, unlike the older code.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>