tabled write_pid_file: unlink before closing
The failure scenario I'm considering here is as follows:

Process A owns the file
Process A closes the file
Process B opens the file, and locks it
Process A unlinks the file (and exits)
Process C opens the file and locks it
Now we have B and C running simultaneously, which may be bad if
one of them is not CLD (e.g. cldbadm or something like that and
so the conflict cannot be detected thanks for shared network ports).

I think that unlinking before closing (and thus unlocking) makes
the scheme perfect, or at any rate I haven't formulated another
failure scenario yet.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>