tabled: Add CLD client
Add a CLD client. It goes through all motions of registering with CLD
and creates necessary files, but there's no lasting effect on operations.
That will come with TDB replication starts using the CLD.

This version does not add a -T flag, but instead starts a private version
of CLD when building. Looking at it from another angle, from now on it's
impossible to build tabled unless a working CLD binary and libraries are
around. This necessitates adding cld to BuildRequires, of course.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>