tabled: Add db4 replication
Honestly, this is not thoroughly tested, just builds and starts enough
for "make distcheck" to pass. In other words, this is a stage for work
in progress. In particular, we need to return correct errors when server
is not a master, and have clients handle it right. Currently, this is
stubbed by not enabling listening until TDB goes into master state.

Note that tdbadm is being pulled tighter and tighter into the orbit
of the running tabled. With this patch, it starts participating in
the replication. It cries for being rearchitected.

Other than the above, it's a milestone for replication in tabled.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>