tabled: return (bogus) Redirect from a slave
The code with replication enabled returned AccessDenied if client
talked to slave server. This is not ideal, since clients do not
retry AccessDenied. The right thing to do is to return a Redirect.

This is not a complete fix, because we need to return URI in
Location: field of a 307 Redirect. Note that it's correct to
return the URI used to access us (unless a proxy really screws
the request). We only stop short because of a need to test this.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>