tabled: use correct name for zone attribute
We're such a young project and we have legacy arrangements already.
In particular what's used to be called a "building" is now a "zone",
for the benefit of users running their clouds out of missile silos,
huts, tents, vehicles, etc.

Jokes aside, the patch makes tabled to use the correct name that
corresponds to what Chunk publishes.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>