tabled: flip the switch to CLD-based configuration
This patch switches tabled from using <StorageNode> clause to reading
what Chunk posted into CLD. This is the default mode used in clouds,
so we want to test it during the build.

Note that tabled should boot normally and tolerate a delayed start-up
of Chunk in either mode, before and after the switch. This patch
demonstrates how to flip the two configuration files atomically.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>