tabled: auto ports, except replication v2
This patch adds the automatic listening port feature to tabled in the
same way it was done in CLD and Chunk. It turns auto ports on, too.

The patch also starts making use of auto ports in CLD and Chunk.
Thus, our build requires this support in the installed versions of them.

There's still an issue of the DB4 replication using a fixed port.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>