tabled: write objects to multiply chunks
This is a small, but self-contained step toward the data redundancy.
The patch adds ability to write data to several Chunk servers.
Along the way, it fixes things up so we use asynchronous writes.

N.B. This patch changes the format of tabled's database in order to
reflect that object IDs are global in the cell. This is done so the
consistency checking and redundancy monitoring and recovery could
be run without accessing the tabled's database. In orther words,
it's a flag day, the database must be re-created.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>