tabled: fix objid when less than 500
When we start over an empty directory, we start with ID 1 in memory and
assume that eventually it will be written out and all will be ok.
Unfortunately, it's wrong. If tabled is interrupted before it created
500 objects, the counter is not written and so objects get confused.
So, force-write the new counter.

I also roll an API change to objid_xxx that I carried for a while.
It permits the objid_xxx to be used where tabled.h is not included,
for example in tdbadm.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>