[test] Stagger the start-daemon
My rule of thumb is that magic delays are evil or stupid, so I worked
on eliminating them from our scripts. However, in this case it's just
not worth it, because the result is that we have to wait way more than
100s for several cycles of CLD timeouts to complete, not just one,
before we declare a failure.

With this change, waiting a bit after CLD is started, all builds
completed that I submitted to Fedora build system.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>