Listen on status earlier
This way status page is useful to show the database status even
when not a master (e.g. a slave).

Unfortunately it's still late to show what's going on while we
bootstrap the CLD client, because with ncld it's synchronous, but
we need the main loop to process events before HTTP service starts
working. But this fix is better than nothing.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>