Fix the flood of "read obj(N)"
Sometimes during operation tabled starts flooding a message "read obj".

I thought it was something about replication thread fetching incomplete
objects, but actually it is caused by attempts to read from a closed
object. It is purely dictated by load levels.

This fix closes the obvious avenue, but leaves any "potential" ones.
I want to see if they ever hit.

Also, fix up messages a bit. In a redundant environment it is necessary
to know NIDs so that a broken Chunk can be isolated, so add that.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>